+++ user_management/models/tests/test_models.py
@@ -3,22 +3,23 @@
 import unittest
 
 import django
-from django.contrib.auth.tokens import default_token_generator
 from django.contrib.sites.models import Site
 from django.core import checks
 from django.db.models import TextField
 from django.db.utils import IntegrityError
 from django.test import TestCase
 from django.utils import six, timezone
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode
-from mock import patch
+from mock import Mock, patch
 
 from user_management.models.tests import utils
+from user_management.utils.notifications import email_context
 from . import models
 from .factories import UserFactory
 from .. import mixins
 
+
+EMAIL_CONTEXT = 'user_management.utils.notifications.email_context'
+SEND_METHOD = 'user_management.utils.notifications.incuna_mail.send'
 
 skip_if_checks_unavailable = unittest.skipIf(
     django.VERSION < (1, 7),
@@ -176,95 +177,49 @@ class TestVerifyEmailMixin(TestCase):
         self.assertFalse(user.is_active)
         self.assertTrue(user.email_verification_required)
 
-    def test_get_email_subject(self):
-        domain = 'http://example.com'
-        expected_subject = '{} account validate'.format(domain)
-        user = self.model()
-
-        self.assertEqual(user.get_email_subject(domain), expected_subject)
-
-    def test_get_email_subject_custom_template(self):
-        domain = 'http://example.com'
-        subject_template = '{domain} register'
-        expected_subject = subject_template.format(domain=domain)
-        user = self.model()
-        user.EMAIL_SUBJECT = subject_template
-
-        self.assertEqual(user.get_email_subject(domain), expected_subject)
-
     def test_email_context(self):
-        site = Site.objects.get_current()
-        user = self.model()
-        uid = urlsafe_base64_encode(force_bytes(user.pk))
+        """Assert `email_context` returns the correct data."""
+        mocked_user = Mock()
+        mocked_site = Mock()
 
-        with patch.object(default_token_generator, 'make_token') as make_token:
-            context = user.email_context(site)
+        class DummyNotification:
+            user = mocked_user
+            site = mocked_site
 
-        expected_context = {'uid': uid, 'token': make_token(), 'site': site}
+        notification = DummyNotification()
+        context = email_context(notification)
+
+        expected_context = {
+            'protocol': 'https',
+            'uid': mocked_user.generate_uid(),
+            'token': mocked_user.generate_token(),
+            'site': mocked_site,
+        }
         self.assertEqual(context, expected_context)
-
-    def test_email_kwargs_default(self):
-        context = {}
-        domain = 'http://example.com'
-        subject = '{} account validate'.format(domain)
-        text_template = 'user_management/account_validation_email.txt'
-        html_template = 'user_management/account_validation_email.html'
-
-        user = self.model(email='dummy@example.com')
-
-        kwargs = user.email_kwargs(context, domain)
-
-        expected_kwargs = {
-            'to': [user.email],
-            'template_name': text_template,
-            'html_template_name': html_template,
-            'subject': subject,
-            'context': context,
-        }
-        self.assertEqual(kwargs, expected_kwargs)
-
-    def test_email_kwargs_custom(self):
-        context = {}
-        domain = 'http://example.com'
-        subject = '{} account validate'.format(domain)
-        text_template = 'user_management/validation_email.txt'
-        html_template = 'user_management/validation_email.html'
-
-        user = self.model(email='dummy@example.com')
-        user.TEXT_EMAIL_TEMPLATE = text_template
-        user.HTML_EMAIL_TEMPLATE = html_template
-
-        kwargs = user.email_kwargs(context, domain)
-
-        expected_kwargs = {
-            'to': [user.email],
-            'template_name': text_template,
-            'html_template_name': html_template,
-            'subject': subject,
-            'context': context,
-        }
-        self.assertEqual(kwargs, expected_kwargs)
 
     def test_send_validation_email(self):
         context = {}
-        kwargs = {'context': context}
         site = Site.objects.get_current()
-        user = self.model()
+        user = self.model(email='email@email.email')
 
-        with patch.object(user, 'email_context') as get_context:
+        with patch(EMAIL_CONTEXT) as get_context:
             get_context.return_value = context
-            with patch.object(user, 'email_kwargs') as get_kwargs:
-                get_kwargs.return_value = kwargs
-                with patch('user_management.models.mixins.send') as send:
-                    user.send_validation_email()
+            with patch(SEND_METHOD) as send:
+                user.send_validation_email()
 
-        get_kwargs.assert_called_once_with(context, site.domain)
-        send.assert_called_once_with(**kwargs)
+        expected = {
+            'to': user.email,
+            'template_name': 'user_management/account_validation_email.txt',
+            'html_template_name': 'user_management/account_validation_email.html',
+            'subject': '{} account validate'.format(site.domain),
+            'context': context
+        }
+        send.assert_called_once_with(**expected)
 
     def test_verified_email(self):
         user = self.model(email_verification_required=False)
 
-        with patch('user_management.models.mixins.send') as send:
+        with patch(SEND_METHOD) as send:
             with self.assertRaises(ValueError):
                 user.send_validation_email()
 
@@ -327,3 +282,28 @@ class TestCustomNameUser(utils.APIRequestTestCase):
     def test_manager_check_invalid(self):
         errors = self.model.check()
         self.assertEqual(errors, [])
+
+
+class TestCustomPasswordResetNotification(TestCase):
+    """Assert we can customise the notification to send a password reset."""
+    model = models.CustomVerifyEmailUser
+
+    def test_send_password_reset_email(self):
+        """Assert `text_email_template` and `html_template_name` can be customised."""
+        context = {}
+        site = Site.objects.get_current()
+        user = self.model(email='email@email.email')
+
+        with patch(EMAIL_CONTEXT) as get_context:
+            get_context.return_value = context
+            with patch(SEND_METHOD) as send:
+                user.send_password_reset()
+
+        expected = {
+            'to': user.email,
+            'template_name': 'my_custom_email.txt',
+            'html_template_name': None,
+            'subject': '{} password reset'.format(site.domain),
+            'context': context
+        }
+        send.assert_called_once_with(**expected)