+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v8.0.1
+
+* Fix translation for notifications.
+
+## v8.0.0
+
+* Use `incuna-pigeon` for notifications.
+
 ## v7.0.1
 
 * Fix `UserChangeForm` admin form `fields` to only include fields used in `UserAdmin.fieldsets`.
@@ -9,7 +17,7 @@
 ### Notes
 
 * When an object is referencing the user model with a foreign key it is possible
-to define the bahavior with `on_delete`.
+to define the behavior with `on_delete`.
 
 see https://docs.djangoproject.com/en/1.7/ref/models/fields/#django.db.models.ForeignKey.on_delete